requirements.txt
Django==1.11.20
django-haystack==2.8.1
django-oscar==1.6.7
#django-extensions==1.7.4
django-widget-tweaks==1.4.3
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==4.1.2
requests==2.21.0
python-dateutil==1.5
#pandas-0.22.0-cp27-cp27m-win32.whl
psycopg2==2.8.2
#django-debug-toolbar==1.9.1